modernize-your-code/codefiles/requirements.txt
streamlit==1.30.0
openai==1.12.0
azure-cosmos==4.5.1
azure-identity==1.15.0
sqlparse==0.4.4
pyodbc==5.0.1
pandas==2.1.4
python-dotenv==1.0.0
